- Change type of OCSQLiteOpenFlags to int to match SQLite's type

+++ ownCloudSDK/Vaults/Database/SQLite/OCSQLiteDB.h
@@ -29,7 +29,7 @@
 @class OCSQLiteQuery;
 @class OCSQLiteTableSchema;
 
-typedef NS_ENUM(NSUInteger, OCSQLiteOpenFlags)
+typedef NS_ENUM(int, OCSQLiteOpenFlags)
 {
 	OCSQLiteOpenFlagsReadOnly = SQLITE_OPEN_READONLY,
 	OCSQLiteOpenFlagsReadWrite = SQLITE_OPEN_READWRITE,
